Avance en el labo

+++ solucion/src/temperature_c.c
@@ -20,13 +20,13 @@ void temperature_c    (
 	unsigned char (*src_matrix)[src_row_size] = (unsigned char (*)[src_row_size]) src;
 	unsigned char (*dst_matrix)[dst_row_size] = (unsigned char (*)[dst_row_size]) dst;
 	int suma = 0;
-	double divido = 0;
+	int divido = 0;
 	for (int i_d = 0, i_s = 0; i_d < filas; i_d++, i_s++) {
 		for (int j_d = 0, j_s = 0; j_d < cols; j_d++, j_s++) {
 			rgb_t *p_d = (rgb_t*)&dst_matrix[i_d][j_d*3];
 			rgb_t *p_s = (rgb_t*)&src_matrix[i_d][j_d*3];
 			suma = p_s->r + p_s->b + p_s->g;
-			divido = (double)suma/3;
+			divido = suma/3;
 			if (between(divido,0,31)){
 				p_d->r = 0;
 				p_d->g =0;
